+++ extensions/veeam.extension/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 1.1
+
+- Added required parameter and conditional parameter dependency validation to `Invoke-PackageParameterValidation`
+- Added `ZeroOneOrTwo` parameter validation
+
 ## 1.0
 
 - Added `Invoke-PackageParameterValidation`